refactor: (#685) 상수 파일 정리
