+++ frontend/__test__/hooks/query/usePostList.test.tsx
@@ -5,7 +5,8 @@ import { renderHook, waitFor } from '@testing-library/react';
 
 import { usePostList } from '@hooks/query/usePostList';
 
-import { POST_TYPE, SORTING, STATUS } from '@constants/post';
+import { POST_TYPE } from '@constants/api';
+import { SORTING, STATUS } from '@constants/post';
 import { QUERY_KEY } from '@constants/queryKey';
 
 import { MOCK_TRANSFORM_GUEST_POST_LIST, MOCK_TRANSFORM_POST_LIST } from '@mocks/mockData/post';
