+++ frontend/src/components/PostForm/validation.ts
@@ -4,6 +4,14 @@ import { WritingVoteOptionType } from '@hooks/useWritingOption';
 
 import { Option } from '@components/common/MultiSelect/types';
 
+import {
+  POST_CATEGORY_POLICY,
+  POST_CONTENT_POLICY,
+  POST_DEADLINE_POLICY,
+  POST_OPTION_POLICY,
+  POST_TITLE_POLICY,
+} from '@constants/policyMessage';
+
 export const checkValidationPost = (
   categoryList: Option[],
   title: string,
@@ -11,17 +19,17 @@ export const checkValidationPost = (
   optionList: WritingVoteOptionType[],
   time: Time
 ) => {
-  if (categoryList.length < 1) return '카테고리를 최소 1개 골라주세요.';
-  if (categoryList.length > 3) return '카테고리를 최대 3개 골라주세요.';
+  if (categoryList.length < 1) return POST_CATEGORY_POLICY.MIN;
+  if (categoryList.length > 3) return POST_CATEGORY_POLICY.MAX;
 
-  if (title.trim() === '') return '제목은 필수로 입력해야 합니다.';
+  if (title.trim() === '') return POST_TITLE_POLICY.REQUIRED;
 
-  if (content.trim() === '') return '내용은 필수로 입력해야 합니다.';
+  if (content.trim() === '') return POST_CONTENT_POLICY.REQUIRED;
 
-  if (optionList.length < 2) return '선택지는 최소 2개 입력해주세요.';
-  if (optionList.length > 5) return '선택지는 최대 5개 입력할 수 있습니다.';
-  if (optionList.some(option => option.text.trim() === '')) return '선택지에 글을 입력해주세요.';
+  if (optionList.length < 2) return POST_OPTION_POLICY.MIN;
+  if (optionList.length > 5) return POST_OPTION_POLICY.MAX;
+  if (optionList.some(option => option.text.trim() === '')) return POST_OPTION_POLICY.REQUIRED;
 
   const isTimeOptionZero = Object.values(time).reduce((a, b) => a + b, 0) < 1;
-  if (isTimeOptionZero) return '시간은 필수로 입력해야 합니다.';
+  if (isTimeOptionZero) return POST_DEADLINE_POLICY.REQUIRED;
 };
