+++ frontend/src/components/comment/CommentList/CommentTextForm/index.tsx
@@ -11,7 +11,7 @@ import { useToast } from '@hooks/useToast';
 import SquareButton from '@components/common/SquareButton';
 import Toast from '@components/common/Toast';
 
-import { COMMENT } from '@constants/comment';
+import { POST_COMMENT } from '@constants/policy';
 
 import { deleteOverlappingNewLine } from '@utils/post/deleteOverlappingNewLine';
 
@@ -97,7 +97,7 @@ export default function CommentTextForm({
         aria-label={isEdit ? '댓글 수정' : '댓글 작성'}
         value={content}
         placeholder="댓글을 입력해주세요. &#13;&#10;타인의 권리를 침해하거나 도배성/광고성/음란성 내용을 포함하는 경우, 댓글의 운영 원칙 및 관련 법률에 의하여 제재를 받을 수 있습니다."
-        onChange={(e: ChangeEvent<HTMLTextAreaElement>) => handleTextChange(e, COMMENT)}
+        onChange={(e: ChangeEvent<HTMLTextAreaElement>) => handleTextChange(e, POST_COMMENT)}
       />
       <S.ButtonContainer>
         {isEdit && (
