+++ frontend/src/components/common/MultiSelect/MultiSelect.stories.tsx
@@ -2,6 +2,10 @@ import type { Meta } from '@storybook/react';
 
 import { useMultiSelect } from '@hooks/useMultiSelect';
 
+import { POST_CATEGORY } from '@constants/policy';
+
+import { MOCK_CATEGORY_LIST, MOCK_OPTION_CATEGORY_LIST } from '@mocks/mockData/categoryList';
+
 import MultiSelect from '.';
 
 const meta: Meta<typeof MultiSelect> = {
@@ -11,39 +15,13 @@ const meta: Meta<typeof MultiSelect> = {
 
 export default meta;
 
-const MOCK_OPTION_LIST = [
-  { id: 1, name: '옵션1' },
-  { id: 2, name: '옵션2' },
-  { id: 5, name: '옵션3' },
-  { id: 7, name: '옵션4' },
-  { id: 9, name: '옵션5' },
-  { id: 10, name: '옵션6' },
-  { id: 11, name: '옵션7' },
-  { id: 13, name: '옵션8' },
-  { id: 15, name: '옵션9' },
-  { id: 16, name: '매우 긴----------~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~옵션10' },
-];
-
-const MOCK_CATEGORY_LIST = [
-  { id: 1, name: '음식' },
-  { id: 2, name: '패션' },
-  { id: 5, name: '금융' },
-  { id: 7, name: '게임' },
-  { id: 9, name: '개발' },
-  { id: 10, name: '연애' },
-  { id: 11, name: '취미' },
-  { id: 13, name: '주식' },
-  { id: 15, name: '연예' },
-  { id: 16, name: '정치' },
-];
-
 export const NotSelected = () => {
   const { selectedOptionList, handleOptionAdd, handleOptionDelete } = useMultiSelect([]);
 
   return (
     <MultiSelect
       selectedOptionList={selectedOptionList}
-      optionList={MOCK_OPTION_LIST}
+      optionList={MOCK_OPTION_CATEGORY_LIST}
       handleOptionAdd={handleOptionAdd}
       handleOptionDelete={handleOptionDelete}
       placeholder="여러 개의 옵션을 선택해주세요."
@@ -52,7 +30,7 @@ export const NotSelected = () => {
 };
 
 export const Selected = () => {
-  const initialSelectedOptionList = MOCK_OPTION_LIST.filter(
+  const initialSelectedOptionList = MOCK_OPTION_CATEGORY_LIST.filter(
     option => option.name === '옵션1' || option.name === '옵션7'
   );
 
@@ -62,7 +40,7 @@ export const Selected = () => {
   return (
     <MultiSelect
       selectedOptionList={selectedOptionList}
-      optionList={MOCK_OPTION_LIST}
+      optionList={MOCK_OPTION_CATEGORY_LIST}
       handleOptionAdd={handleOptionAdd}
       handleOptionDelete={handleOptionDelete}
     />
@@ -70,11 +48,9 @@ export const Selected = () => {
 };
 
 export const CategoryNotSelected = () => {
-  const CATEGORY_COUNT_LIMIT = 3;
-
   const { selectedOptionList, handleOptionAdd, handleOptionDelete } = useMultiSelect(
     [],
-    CATEGORY_COUNT_LIMIT
+    POST_CATEGORY.MAX_AMOUNT
   );
 
   return (
