+++ frontend/src/constants/animation.ts
@@ -1,2 +1,4 @@
-//초단위
+/**
+ * 토스트가 사용자에게 보여지는 시간을 뜻 합니다.
+ */
 export const TOAST_TIME = 3;
