--- frontend/src/constants/announcement.ts
@@ -1,5 +1,0 @@
-export const APP_LAUNCH_EVENT = {
-  TITLE: '🎉보투게더 출시 기념 이벤트 진행 중',
-  CONTENT: '이벤트 참여하면 상품을 받을 수 있다고?',
-  CONTENT_DETAIL: '',
-};
