--- frontend/src/constants/comment.ts
@@ -1,4 +1,0 @@
-export const COMMENT = {
-  MAX_LENGTH: 200,
-  MIN_LENGTH: 1,
-} as const;
