--- frontend/src/constants/vote.ts
@@ -1,3 +1,0 @@
-export const POST = {
-  NOT_VOTE: 0,
-};
