+++ frontend/src/hooks/useCurrentKeyword.ts
@@ -1,17 +1,16 @@
 import { useSearchParams } from 'react-router-dom';
 
-import { DEFAULT_KEYWORD, SEARCH_KEYWORD, SEARCH_KEYWORD_MAX_LENGTH } from '@constants/post';
+import { SEARCH_KEYWORD } from '@constants/api';
+import { SEARCH_KEYWORD_MAX_LENGTH } from '@constants/policy';
 
 import { getTrimmedWord } from '@utils/getTrimmedWord';
 
 export const useCurrentKeyword = () => {
   const [searchParams] = useSearchParams();
   const currentKeyword =
-    searchParams.get(SEARCH_KEYWORD)?.toString().slice(0, SEARCH_KEYWORD_MAX_LENGTH) ??
-    DEFAULT_KEYWORD;
+    searchParams.get(SEARCH_KEYWORD)?.toString().slice(0, SEARCH_KEYWORD_MAX_LENGTH) ?? '';
 
   return {
-    currentKeyword:
-      currentKeyword !== DEFAULT_KEYWORD ? getTrimmedWord(currentKeyword) : currentKeyword,
+    currentKeyword: currentKeyword !== '' ? getTrimmedWord(currentKeyword) : currentKeyword,
   };
 };
