+++ frontend/src/hooks/useWritingOption.tsx
@@ -1,17 +1,14 @@
 import React, { ChangeEvent, useRef, useState } from 'react';
 
-import { uploadImage } from '@utils/post/uploadImage';
+import { POST_WRITING_OPTION } from '@constants/policy';
 
-const MAX_WRITING_LENGTH = 50;
+import { uploadImage } from '@utils/post/uploadImage';
 
 export interface WritingVoteOptionType {
   id: number;
   text: string;
   imageUrl: string;
 }
-
-const MIN_COUNT = 2;
-const MAX_COUNT = 5;
 
 const INIT_OPTION_LIST = [
   { id: Math.floor(Math.random() * 100000), text: '', imageUrl: '', isServerId: false },
@@ -27,7 +24,7 @@ export const useWritingOption = (initialOptionList?: WritingVoteOptionType[]) =>
   const contentInputRefList = useRef<HTMLInputElement[]>([]);
 
   const addOption = () => {
-    if (optionList.length >= MAX_COUNT) return;
+    if (optionList.length >= POST_WRITING_OPTION.MAX_COUNT) return;
 
     const updatedOptionList = [
       ...optionList,
@@ -42,9 +39,9 @@ export const useWritingOption = (initialOptionList?: WritingVoteOptionType[]) =>
       const { value } = event.target;
       const standard = value.length;
 
-      if (standard === MAX_WRITING_LENGTH) {
+      if (standard === POST_WRITING_OPTION.MAX_LENGTH) {
         event.target.setCustomValidity(
-          `선택지 내용은 ${MAX_WRITING_LENGTH}자까지 입력 가능합니다.`
+          `선택지 내용은 ${POST_WRITING_OPTION.MAX_LENGTH}자까지 입력 가능합니다.`
         );
         event.target.reportValidity();
         return;
@@ -64,7 +61,7 @@ export const useWritingOption = (initialOptionList?: WritingVoteOptionType[]) =>
     };
 
   const deleteOption = (optionId: number) => {
-    if (optionList.length <= MIN_COUNT) return;
+    if (optionList.length <= POST_WRITING_OPTION.MIN_LENGTH) return;
 
     const removedOptionList = optionList.filter(optionItem => optionItem.id !== optionId);
