+++ frontend/src/pages/HomePage/types.ts
@@ -1,8 +1,5 @@
-import {
-  REQUEST_POST_KIND_URL,
-  REQUEST_SORTING_OPTION,
-  REQUEST_STATUS_OPTION,
-} from '@constants/post';
+import { REQUEST_POST_KIND_URL } from '@constants/api';
+import { REQUEST_SORTING_OPTION, REQUEST_STATUS_OPTION } from '@constants/post';
 
 export type PostStatus = keyof typeof REQUEST_STATUS_OPTION;
 export type PostSorting = keyof typeof REQUEST_SORTING_OPTION;
