+++ frontend/src/types/report.ts
@@ -1,4 +1,4 @@
-import { REPORT_MESSAGE } from '@components/ReportModal/constants';
+import { REPORT_MESSAGE } from '@constants/policyMessage';
 
 export type ReportType = 'POST' | 'COMMENT' | 'NICKNAME';
